+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -25,6 +25,10 @@ Unreleased
 3.15.0 - (April 24, 2019)
 ------------------
 ### Added
+* Added timeVariant prop to datetime picker to support 12-hour format.
+* Added example for time variant on terra-dev site
+
+### Added
 * Added onSelect function prop that fires when selecting a date via date picker.
 
 ### Changed